backend/requirements.txt
# Django Backend Dependencies for Sernion Mark

# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-oauth-toolkit==2.2.0

# Database
django-extensions==3.2.3
psycopg2-binary==2.9.9

# File Handling & Media
Pillow==10.1.0
python-magic==0.4.27
django-storages==1.14.2
boto3==1.34.0

# API Documentation
drf-yasg==1.21.7

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Utilities
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0

# Rate Limiting
django-ratelimit==4.1.0

# Email
django-anymail==10.2

# Environment Management
django-environ==0.11.2
